Reorganized code in preparation for more files.
