+++ wasm/src/NumericMatrix.h
@@ -1,21 +1,10 @@
-#include <emscripten/bind.h>
-#include "tatami/base/DenseMatrix.hpp"
+#ifndef NUMERIC_MATRIX_H
+#define NUMERIC_MATRIX_H
 
-template<typename T>
-class JSVector {
-public:
-    JSVector(const T* p, size_t n) : ptr(p), num(n) {}
-    size_t size() const { return num; }
-    const T* data() const { return ptr; }
-    const T* begin() const { return ptr; }
-    const T* end() const { return ptr + num; }
-private:
-    const T* ptr;
-    size_t num;
-};
+#include "JSVector.h"
+#include "tatami/tatami.h"
 
-class NumericMatrix {
-public:
+struct NumericMatrix {
     NumericMatrix(int nr, int nc, uintptr_t values) {
         JSVector<double> thing(reinterpret_cast<const double*>(values), nr*nc);
         ptr = std::shared_ptr<tatami::numeric_matrix>(new tatami::DenseRowMatrix<double, int, decltype(thing)>(nr, nc, thing));
@@ -48,16 +37,7 @@ public:
         return;
     }
 
-private:
     std::shared_ptr<tatami::numeric_matrix> ptr;
 };
 
-EMSCRIPTEN_BINDINGS(my_class_example) {
-    emscripten::class_<NumericMatrix>("NumericMatrix")
-        .constructor<int, int, uintptr_t>()
-        .function("nrow", &NumericMatrix::nrow)
-        .function("ncol", &NumericMatrix::ncol)
-        .function("row", &NumericMatrix::row)
-        .function("column", &NumericMatrix::column)
-        ;
-}
+#endif
